requirements.txt
Flask==3.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
APScheduler==3.10.4
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
